Create csv migration management command
I've created a management command that'll read from the csv file and migrate the data to the db. I'm assuming the command will come in handy later. However, if migration from a csv file is a one time thing and only at the start of the project, it could've also been a simple Django data migration.

After running migrations, to migrate data from the csv file all you need to do is:

```
$ python manage.py migrate_data_from_csv
```

Note: The reason there are multiple migration files when only 2 would've been enough is, my slow discovery of null values in the CSV. I was assuming not everything can be blank/null